Refactored the DAO tests.
The ApplicationDaoTest had some methods to test if the modified field is working correctly. As this field is will be updated in the GenericHibernateDao, this behaviour should be tested in a test for the GenericHibernateDao (and not in the ApplicationDaoTest). For this reason a new test "GenericHibnernateDaoTest" has been created. As GenericHibernateDao is an abstract class, we cannot instantiate it to test this. Hence the ApplicationDao (as an simple extension of the GenericHibernateDao) is used in the new test.
